go back to using classic confinement for snap
Required plugs for this to work do not currently exist in snapd.
I'll leave this here for now in case they are added in the future
or if anyone wants to build their own snap with classic confinement.

Only a few of the plugs this commit removes were actually required.
I don't think there are any other currently existing plugs that
are required for this to work.